remove functionality of 'select_function()'
It was a provisional to commit functions as argument
before OpenSCAD version 2021.01.

You could pass a name for the function.
The function 'select_function()' has searched this name for a hit
and called the written function with its arguments.

You had to write the function call with the test of the name into
the internal function 'select_function()' or to define a
function 'select_extern()' to do this.
I think this is not good for code maintenance.
It was slow.
And with the new function literals since OpenSCAD version 2021.01
it is unnecessary. Its a dinosaur of code.