unify parameter 'slices' to control fragments of a curve
If slices is set, the object will always get this count of fragments.
E.g. a segment of a circle defined with 'angle',
this segment gets this count of fragments.
If you want to get the behavior like 'rotate_extrude()' to split the
count of fragments to the segment part and '$fn' is not useful,
you can use the helper functions to calculate this e.g.
- get_slices_circle_current   (r, angle)
- get_slices_circle_current_x (r, angle)

update documentation
- most functions and modules had already this behavior

update modules to the correct behavior of 'slices':
- cylinder_edges_fillet()
- edge_ring_fillet()

add argument 'slices' to module 'rotate_extrude_extend()'

add helper functions:
- parameter_slices_circle()
- parameter_slices_circle_x()
- get_fn_circle()

remove unused parameter 'r' in internal function 'get_slices_circle_pie()'